Fix packer json creation
Previously this code had a hack in which it would get the packer config
as a string and then lazily echo that into a remote file. This should
be done with a local temporary directory and a put instead. Oh and make
the json look pretty while we are at it.